requirements.txt
asgiref==3.5.2
astroid==2.11.5
black==22.3.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.5.1
Django==4.0.4
djangorestframework==3.13.1
Faker==13.11.1
idna==3.3
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pylint==2.13.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.2.0
ujson==5.3.0
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.1
